% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circan_square.R
\name{arccot}
\alias{arccot}
\title{arccot}
\usage{
arccot(x)
}
\arguments{
\item{x}{integer to compute the arc cotangent on.}
}
\description{
Arc cotangent function. Auxiliary to the square wave function.
}
\examples{
arccot(x=2pi)
}
\keyword{arccot}
