% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{best_fit_select}
\alias{best_fit_select}
\title{best_fit_select}
\usage{
best_fit_select(results_list)
}
\arguments{
\item{results_list}{A list of datasets with different regression results}
}
\description{
Obtains the curve form from a series of fit results that fit best each feature.
}
\examples{
results_list <- list.files(path="~/results/", pattern="*.csv")
best_fit_results <- do.call("best_fit_select", as.list(results_list))
}
\keyword{NA,}
\keyword{regression}
