% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircaN_library.R
\name{circan_groups}
\alias{circan_groups}
\title{circan_groups}
\usage{
circan_groups(
  data,
  s2c,
  id_col,
  sample_col,
  time_col,
  group_col,
  ind_col,
  shiny = FALSE,
  mode = "default",
  init_value = 24,
  max_per = Inf,
  min_per = -Inf
)
}
\arguments{
\item{data}{Dataframe containing the expression data of samples from both groups. Samples must be in columns and genes in rows.}

\item{s2c}{Dataframe containing the metadata for the samples. Must have at least a 'sample' column
with the sample name as it appears in the data matrix;
a 'time' column with the time point the sample was collected;
and an 'ind' column containing information for the individual the sample comes from.}

\item{id_col}{Column in the expression dataframe containing the feature IDs.}

\item{sample_col}{Column in the metadata dataframe containing the names of the samples.}

\item{time_col}{Column in the metadata dataframe containing the time point of each sample.}

\item{group_col}{Column in the metadata dataframe containing the group to which sample belongs.}

\item{ind_col}{Column in the metadata dataframe containing the individual from which each sample comes.
Please note that the individuals from both groups must be different.}

\item{shiny}{Is the package running in a shiny app? default to FALSE.}

\item{mode}{Algorithm to use in the NLS regression. Must be one of 'default' for Gauss-Newton, 'plinear' for the Golub-Pereyra algorithm
for partially linear least-squares models and 'port' for the ‘nl2sol’ algorithm from the Port library. Default is default. See nls documentation
for extended info.}

\item{init_value}{Initial value for the period. Default is set to 24.}

\item{max_per}{Maximum period to regress. Default is set to Inf.}

\item{min_per}{Minimum period to regress. Default is set to -Inf.}
}
\description{
Nonlinear least squares model for accurate detection of circadian expression patterns and
group comparison.
}
\keyword{CircaN}
\keyword{circadian}
\keyword{comparison.}
\keyword{group}
\keyword{regression,}
