% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{data_summary}
\alias{data_summary}
\title{data_summary}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{dataframe with data to summarize}

\item{varname}{column name for the variable which we want to summarize}

\item{groupnames}{grouping variables}
}
\description{
Obtains the sd and mean grouping by column. Will return a df with the mean and sd by group of
the chosen variable.
from: STHDA barplot tutorial
}
\examples{
results_list <- data_summary(ToothGrowth, varname="len", groupnames=c("supp", "dose"))
summarized_data <- data_summary(all_data, varname = "meta2d_AMP", groupnames = "amp")
}
\keyword{mean,}
\keyword{sd,}
\keyword{summarize}
