% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fisher_method}
\alias{fisher_method}
\title{fisher_method}
\usage{
fisher_method(x)
}
\arguments{
\item{x}{numeric vector with the p.values to merge}
}
\description{
Gets combined p.val by Fisher's method
}
\keyword{fisher}
\keyword{p.value}
