% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircaN_library.R
\name{full_mode_analysis}
\alias{full_mode_analysis}
\title{full_mode_analysis}
\usage{
full_mode_analysis(
  data,
  s2c,
  algorithms = c("circan", "jtk", "metacycle"),
  min_per = 20,
  max_per = 28,
  circan_init_value = 24,
  circan_mode = "port",
  mc_cycMethod = c("LS", "JTK")
)
}
\arguments{
\item{data}{The data matrix with the omics data.}

\item{s2c}{Data frame with at least the columns sample (which must coincide with the samples in data) and time (recording
time corresponding to each sample.}

\item{algorithms}{character vector indicating which algorithms to test the data with. Must beone, or  a combination
of circan, jtk and metacycle.}

\item{min_per}{Minimum period to search for.}

\item{max_per}{Maximum period to search for.}

\item{circan_init_value}{Initial value for the period of CircaN regression.}

\item{circan_mode}{Algorithm to use in CircaN regression. Must be one of 'default' for Gauss-Newton, 'plinear' for the Golub-Pereyra algorithm
for partially linear least-squares models and 'port' for the ‘nl2sol’ algorithm from the Port library.}

\item{mc_cycMethod}{a character vector(length 1 or 2 or 3) for MetaCycle. User-defined methods for detecting rhythmic signals, must be
selected as any one, any two or all three methods(default) from "ARS"(ARSER), "JTK"(JTK_CYCLE) and "LS"(Lomb-Scargle).}
}
\description{
Runs analysis on CircaN, JTK and MetaCycle and integrates the results.
}
\examples{
results <- full_mode_analysis(my_data, s2c)
}
\keyword{circan,}
\keyword{jtk,}
\keyword{metacycle}
