% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_fisher}
\alias{get_fisher}
\title{get_fisher}
\usage{
get_fisher(...)
}
\arguments{
\item{...}{numeric vectors with the p.values to merge}
}
\description{
Uses fisher_method to calculate the fisher combined p.value of the provided numeric vectors
in a pairwise manner.
}
\examples{
circan_results$combined_pval <- get_fisher(as.numeric(circan_results$p.value.amp), as.numeric(circan_results$p.value.per))
}
\keyword{fisher}
\keyword{p.value}
