% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{meta2d_wrapper}
\alias{meta2d_wrapper}
\title{meta2d_wrapper}
\usage{
meta2d_wrapper(
  data,
  timepoints,
  min_per = 20,
  max_per = 28,
  cycMethod = c("LS", "JTK")
)
}
\arguments{
\item{data}{The data matrix with the omics data.}

\item{timepoints}{numeric vector containing the timepoints corresponding to each column in data.}

\item{cycMethod}{a character vector(length 1 or 2 or 3). User-defined methods for detecting rhythmic signals, must be
selected as any one, any two or all three methods(default) from "ARS"(ARSER), "JTK"(JTK_CYCLE) and "LS"(Lomb-Scargle).}
}
\description{
Wraps the meta 2d method with JTK + LS and returns a data frame with the results. The period is set between 20 and 28h.
}
\examples{
meta2d_results <- meta2d_wrapper(my_data, s2c$time)
}
\keyword{meta2d}
\keyword{metacycle,}
