% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{statistics_calc}
\alias{statistics_calc}
\title{statistics_calc}
\usage{
statistics_calc(detected_genes, real_positive, real_negative)
}
\arguments{
\item{detected_genes}{the genes detected as positives}

\item{real_positive}{real circadian genes}

\item{real_negative}{real non-circadian genes}
}
\description{
Obtains the number of tp, fp, fn and tn contained in a dataset
}
\examples{
results_list <- list.files(path="~/results/", pattern="*.csv")
best_fit_results <- do.call("best_fit_select", as.list(results_list))
}
\keyword{regression}
