% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{supress_na_warning}
\alias{supress_na_warning}
\title{supress_na_warning}
\usage{
supress_na_warning(w)
}
\description{
Supresses warning when BH q value is computed on datasets with NAs
}
\examples{
withCallingHandlers(BH <- p.adjust(as.numeric(results$p.value.per), method = "BH"), warning = supress_na_warning)
}
\keyword{NA,}
\keyword{regression}
